fix position:fixed css glitch by separating components
