+++ sites/kit.svelte.dev/src/lib/search/stores.js
@@ -1,0 +1,4 @@
+import { writable } from 'svelte/store';
+
+export const searching = writable(false);
+export const query = writable('');
